refactor(graphql): extract Environments to a separate file
