+++ alchemist-graphql-surrogates/src/main/kotlin/it/unibo/alchemist/boundary/graphql/monitor/EnvironmentSubscriptionMonitor.kt
@@ -51,19 +51,3 @@ class EnvironmentSubscriptionMonitor<T, P : Position<out P>> : OutputMonitor<T, 
         this.stepDone(environment, null, Time.ZERO, 0L)
     }
 }
-
-/**
- * Extension object for [Environment]s providing a function to retrieve the
- * [EnvironmentSubscriptionMonitor] of this [Environment].
- */
-object Environments {
-    /**
-     * Returns the [EnvironmentSubscriptionMonitor] of this [Environment].
-     * @return the [EnvironmentSubscriptionMonitor] of this [Environment]
-     */
-    fun <T, P>Environment<T, P>.subscriptionMonitor(): EnvironmentSubscriptionMonitor<T, P>
-        where P : Position<out P> =
-        this.simulation.outputMonitors.filterIsInstance<EnvironmentSubscriptionMonitor<T, P>>()
-            .apply { check(size == 1) { "Only one subscription monitor is allowed" } }
-            .first()
-}
